Generate `"Primary Energy"` CSV header programmatically
This was the only line that was left longer than 79 characters after
reformatting the file. While I could have broken this up into multiple
lines using standard string concatenation, I chose to generate the
header programmatically. This makes it easier to change the width, the
label or the fill character and makes the code way shorter. It changes
the header to now be centered, but I hope this isn't a problem.
For those who want know that the new line means, the curly braces just
tell Python to replace them and their contents, the colon tells Python
that everything after it specifies how the interpolated string should be
formatted and the other characters have the following meaning:

  - `=`: is the character used to fill empty space in the interpolated
    string,
  - `^`: is the alignment specifier telling Python to center the
    interpolated string and
  - `73`: tells Python that the interpolated string should be 73
    characters long.